fix: Rename pages and components
delete quentin.jsx
