+++ src/components/App.jsx
@@ -14,8 +14,8 @@ import { translate } from 'cozy-ui/transpiled/react/I18n'
 import Devices from 'containers/Devices'
 import IntentRedirect from 'services/IntentRedirect'
 import Passphrase from 'containers/Passphrase'
-import Permission from 'containers/Permission'
-import PermissionsApplication from 'containers/PermissionsApplication'
+import Permission from 'components/Permissions/PermissionDetails'
+import PermissionsApplication from 'components/Permissions/AppPermissions'
 import PermissionsTab from 'components/Permissions/PermissionsTab'
 import Profile from 'containers/Profile'
 import Sessions from 'containers/Sessions'
